+++ nominal/core/_clientsbunch.py
@@ -32,7 +32,7 @@ from nominal_api import (
 )
 from typing_extensions import Self
 
-from nominal.core._utils import create_conjure_client_factory
+from nominal.core._utils.networking import create_conjure_client_factory
 from nominal.ts import IntegralNanosecondsUTC