+++ nominal/core/_stream/batch_processor.py
@@ -5,11 +5,11 @@ from typing import Sequence, cast
 
 from nominal_api import storage_writer_api
 
-from nominal.core.stream.write_stream import BatchItem
+from nominal.core._stream.write_stream import BatchItem, LogItem
 from nominal.ts import _SecondsNanos
 
 
-def make_points(api_batch: Sequence[BatchItem]) -> storage_writer_api.PointsExternal:
+def make_points(api_batch: Sequence[BatchItem[str | float | int]]) -> storage_writer_api.PointsExternal:
     if isinstance(api_batch[0].value, str):
         return storage_writer_api.PointsExternal(
             string=[
@@ -44,7 +44,7 @@ def make_points(api_batch: Sequence[BatchItem]) -> storage_writer_api.PointsExte
 
 
 def process_batch_legacy(
-    batch: Sequence[BatchItem],
+    batch: Sequence[BatchItem[str | float | int]],
     nominal_data_source_rid: str,
     auth_header: str,
     storage_writer: storage_writer_api.NominalChannelWriterService,
@@ -67,3 +67,35 @@ def process_batch_legacy(
         auth_header,
         request,
     )
+
+
+def process_log_batch(
+    batch: Sequence[LogItem],
+    nominal_data_source_rid: str,
+    auth_header: str,
+    storage_writer: storage_writer_api.NominalChannelWriterService,
+) -> None:
+    def _get_channel_name(batch_item: LogItem) -> str:
+        return batch_item.channel_name
+
+    # Not using BatchItem.sort_key, as we don't need to group by tags-- each log
+    # has its own set of args when streamed.
+    batches_by_channel = itertools.groupby(sorted(batch, key=_get_channel_name), key=_get_channel_name)
+    requests = [
+        storage_writer_api.WriteLogsRequest(
+            logs=[
+                storage_writer_api.LogPoint(
+                    timestamp=_SecondsNanos.from_nanoseconds(batch_item.timestamp).to_api(),
+                    value=storage_writer_api.LogValue(
+                        message=batch_item.value,
+                        args={k: v for k, v in (batch_item.tags or {}).items()},
+                    ),
+                )
+                for batch_item in batch_by_channel
+            ],
+            channel=channel,
+        )
+        for channel, batch_by_channel in batches_by_channel
+    ]
+    for request in requests:
+        storage_writer.write_logs(auth_header, nominal_data_source_rid, request)