+++ nominal/core/_utils/pagination_tools.py
@@ -171,6 +171,18 @@ def search_runs_paginated(
         yield from response.results
 
 
+def search_runs_by_asset_paginated(
+    run: scout.RunService,
+    auth_header: str,
+    asset_rid: str,
+) -> Iterable[scout_run_api.Run]:
+    def factory(page_token: str | None) -> scout_run_api.GetRunsByAssetRequest:
+        return scout_run_api.GetRunsByAssetRequest(asset=asset_rid, next_page_token=page_token)
+
+    for response in paginate_rpc(run.get_runs_by_asset, auth_header, request_factory=factory):
+        yield from response.results
+
+
 def search_secrets_paginated(
     secrets: secrets_api.SecretService, auth_header: str, query: secrets_api.SearchSecretsQuery
 ) -> Iterable[secrets_api.Secret]: