+++ nominal/experimental/compute/dsl/__init__.py
@@ -1,0 +1,25 @@
+from nominal.experimental.compute.dsl.exprs import (
+    Expr,
+    NumericExpr,
+    RangeExpr,
+)
+from nominal.experimental.compute.dsl.params import (
+    DoubleConstant,
+    Nanoseconds,
+    NanosecondsUTC,
+    RollingOperationLiteral,
+    ThresholdOperatorLiteral,
+    TimeUnitLiteral,
+)
+
+__all__ = [
+    "Expr",
+    "NumericExpr",
+    "RangeExpr",
+    "Nanoseconds",
+    "NanosecondsUTC",
+    "DoubleConstant",
+    "TimeUnitLiteral",
+    "ThresholdOperatorLiteral",
+    "RollingOperationLiteral",
+]