+++ nominal/experimental/stream_v2/_serializer.py
@@ -5,8 +5,8 @@ from dataclasses import dataclass
 
 from typing_extensions import Self
 
-from nominal.core._utils import Batch
-from nominal.core.stream.batch_processor_proto import SerializedBatch, serialize_batch
+from nominal.core._stream.batch_processor_proto import SerializedBatch, serialize_batch
+from nominal.core._utils.queueing import Batch
 
 
 @dataclass(frozen=True)
@@ -26,7 +26,7 @@ class BatchSerializer:
         pool = ProcessPoolExecutor(max_workers=max_workers)
         return cls(pool=pool)
 
-    def serialize(self, batch: Batch) -> Future[SerializedBatch]:
+    def serialize(self, batch: Batch[str | float]) -> Future[SerializedBatch]:
         return self.pool.submit(serialize_batch, batch)
 
     def __enter__(self) -> BatchSerializer: